+++ src/tests/decider-distro-tests-rs/tests/utils/distro.rs
@@ -1,0 +1,93 @@
+use crate::utils::default::IPFS_MULTIADDR;
+use decider_distro::DeciderConfig;
+use fluence_app_service::TomlMarineConfig;
+use fluence_spell_dtos::trigger_config::TriggerConfig;
+use system_services::{PackageDistro, ServiceDistro, SpellDistro};
+
+pub fn package_items_names(distro: &PackageDistro) -> Vec<String> {
+    distro
+        .services
+        .iter()
+        .map(|s| s.name.clone())
+        .chain(distro.spells.iter().map(|s| s.name.clone()))
+        .collect()
+}
+
+pub fn make_distro(trigger_config: TriggerConfig, settings: DeciderConfig) -> PackageDistro {
+    let connector = decider_distro::connector_service_modules();
+    let marine_config: TomlMarineConfig =
+        toml::from_slice(connector.config).expect("parse marine config");
+    let service = ServiceDistro {
+        modules: connector.modules,
+        config: marine_config,
+        name: connector.name.to_string(),
+    };
+
+    let distro_spell = decider_distro::decider_spell(settings);
+    let spell = SpellDistro {
+        name: "decider".to_string(),
+        air: distro_spell.air,
+        kv: distro_spell.kv.clone(),
+        trigger_config,
+    };
+
+    PackageDistro {
+        name: "decider".to_string(),
+        version: decider_distro::VERSION,
+        services: vec![service],
+        spells: vec![spell],
+        init: None,
+    }
+}
+
+pub fn make_distro_default() -> PackageDistro {
+    let decider_settings = DeciderConfig {
+        worker_period_sec: 0,
+        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
+        chain_api_endpoint: "http://127.0.0.1:12009".to_string(),
+        chain_network_id: 11,
+        chain_contract_block_hex: "0x0".to_string(),
+        chain_matcher_addr: "0x0".to_string(),
+        chain_workers_gas: 210_00,
+        chain_wallet_key: "0x0".to_string(),
+    };
+    // let's try to run a decider cycle on demand by updating the config
+    let mut trigger_config = TriggerConfig::default();
+    trigger_config.clock.start_sec = 1;
+    make_distro(trigger_config, decider_settings)
+}
+
+pub fn make_distro_with_api(api: String) -> PackageDistro {
+    let decider_settings = DeciderConfig {
+        // worker will run once
+        worker_period_sec: 0,
+        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
+        chain_api_endpoint: api,
+        chain_network_id: 11,
+        chain_contract_block_hex: "0x0".to_string(),
+        chain_matcher_addr: "0x0".to_string(),
+        chain_workers_gas: 210_00,
+        chain_wallet_key: "0xfdc4ba94809c7930fe4676b7d845cbf8fa5c1beae8744d959530e5073004cf3f"
+            .to_string(),
+    };
+    // decider will run once
+    let trigger_config = TriggerConfig::default();
+    make_distro(trigger_config, decider_settings)
+}
+
+pub fn make_distro_with_api_and_config(api: String, config: TriggerConfig) -> PackageDistro {
+    let decider_settings = DeciderConfig {
+        // worker will run once
+        worker_period_sec: 0,
+        worker_ipfs_multiaddr: IPFS_MULTIADDR.to_string(),
+        chain_api_endpoint: api,
+        chain_network_id: 11,
+        chain_contract_block_hex: "0x0".to_string(),
+        chain_matcher_addr: "0x0".to_string(),
+        chain_workers_gas: 210_00,
+        chain_wallet_key: "0xfdc4ba94809c7930fe4676b7d845cbf8fa5c1beae8744d959530e5073004cf3f"
+            .to_string(),
+    };
+    // decider will run once
+    make_distro(config, decider_settings)
+}
